+++ NEWS.md
@@ -24,6 +24,8 @@
 
 - `rst2pdf()` will pass a default output filename to `rst2pdf` (if the input is `foo.rst`, the output will be `foo.pdf`)
 
+- `knit2wp()` gained a new argument `publish = TRUE` (thanks, Eric Nantz) (#512)
+
 ## BUG FIXES
 
 - fixed the problem reported in http://stackoverflow.com/q/12448507/559676 now \usepackage{upquote} should appear after \usepackage{fontenc}, so single quotes in verbatim environments will no longer cause problems